refactor: better structure for iOS code
