+++ flutter_readium/ios/flutter_readium/Sources/flutter_readium/utils/EventStreamHandler.swift
